moves current step highlighting functions to utils
