+++ src/components/editor/controls/Row/Row.jsx
@@ -1,0 +1,36 @@
+import { Step } from '../index.js';
+import styles from './Row.css';
+
+const Row = ({ notes, setNotes, keyArray }) => {
+  const handleNoteChange = (e) => {
+    const indexOfStep = e.target.id.split('-')[1];
+
+    const indexOfKeyArray = keyArray.findIndex(
+      (note) => note === e.target.textContent
+    );
+
+    const newNotes = notes.map((note, index) => {
+      if (Number(indexOfStep) === index) {
+        return keyArray[indexOfKeyArray + 1];
+      }
+      return note;
+    });
+
+    setNotes(newNotes);
+  };
+
+  return (
+    <div className={styles.row}>
+      {notes.map((note, index) => (
+        <Step
+          key={`step-${index}`}
+          note={note}
+          index={index}
+          handleNoteChange={handleNoteChange}
+        />
+      ))}
+    </div>
+  );
+};
+
+export default Row;
