+++ src/components/editor/controls/index.js
@@ -7,6 +7,8 @@ import Controls from './TrackControls/TrackControls.jsx';
 import Dropdown from './Dropdown/Dropdown.jsx';
 import Joystick from './Joystick/Joystick.jsx';
 import PlayButton from './PlayButton/PlayButton.jsx';
+import Step from './Step/Step.jsx';
+import Row from './Row/Row.jsx';
 import VolumeFader from './VolumeFader/VolumeFader.jsx';
 import GlobalControls from './GlobalControls/GlobalControls.jsx';
 
@@ -18,6 +20,8 @@ export {
   Dropdown,
   Joystick,
   PlayButton,
+  Step,
+  Row,
   VolumeFader,
   GlobalControls,
 };
